实现用户的登录，注册

+++ src/main/java/com/miaoshaproject/service/model/UserModel.java
@@ -31,7 +31,7 @@ public class UserModel {
         return gender;
     }
 
-    public void setGender(Integer gender) {
+    public void setGender(Byte gender) {
         this.gender = gender;
     }
